Require more explicit permission to run queries from Copilot
