+++ MarkMpn.Sql4Cds.XTB/MarkMpn.Sql4Cds.XTB.csproj
@@ -67,6 +67,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CopilotScriptObject.cs" />
     <Compile Include="CreateCopilotAssistantForm.cs">
       <SubType>Form</SubType>
     </Compile>
